requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
concurrent-log-handler==0.9.19
filelock==3.4.2
Flask==2.0.2
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
portalocker==2.3.2
pymongo==4.0.1
pywin32==303
PyYAML==6.0
requests==2.27.1
requests-file==1.5.1
six==1.16.0
tldextract==3.1.2
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
